data-provider-example/requirements.txt
flask
flask-socketio
websocket-client
eth_utils
